P1 — Money is stored as FLOAT  #170

+++ backend/src/models/Transaction.js
@@ -12,7 +12,7 @@ const Transaction = sequelize.define('Transaction', {
     allowNull: false,
   },
   amount: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(12, 2),
     allowNull: false,
   },
   type: {
